requirements.txt
streamlit>=1.31.0
groq>=0.4.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
beautifulsoup4>=4.12.0
requests>=2.31.0
numpy>=1.24.0
python-dotenv>=1.0.0
lxml>=5.0.0
